Déplace les routes /api/*
